% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotonicity_plot.R
\name{monotonicity_plots}
\alias{monotonicity_plots}
\title{Create monotonicity plots for numeric predictors in a logistic model}
\usage{
monotonicity_plots(df, response, baseline = NULL)
}
\arguments{
\item{df}{a data frame containing the predictors and response for the model}

\item{response}{A character string specifying the response variable with 2 variables.}

\item{baseline}{Optional. A character string specifying the reference
level of the response variable. If supplied, the response factor will
be re-leveled to this baseline.}
}
\value{
\code{ggplot2} objects.
}
\description{
This function fits monotonicity plots for all numeric predictors
by automatically identifying the numeric predictors,
setting a desired baseline level for the response, and converting the
response factor to a 0/1 numeric variable.
}
\details{
Analyzing model diagnostics becomes simpler using this function because it
automates the process of creating monotonicity plots by identifying numeric
variables and producing the monotonicity plots
}
\examples{
\dontrun{
df <- data.frame(
  y = c("no", "yes", "no", "yes", "no", "yes", "no", "yes"),
  x = c(1, 2, 3, 4, 3, 7, 2, 6)
)

monotonicity_plots(df, response = "y", baseline = "no")
}

}
